fix: make environment highlight []
